feat(load_project_dlg): improve error handling and performance

+++ main.py
@@ -302,10 +302,10 @@ class Giswater(QObject):
                 open(filepath, 'a').close()
 
             if not global_vars.gw_dev_mode:
-                # Copy user_params file to user config folder
                 src = f"{lib_vars.plugin_dir}{os.sep}config{os.sep}user_params.config"
                 dst = f"{config_folder}{os.sep}user_params.config"
-                shutil.copyfile(src, dst)
+                if not os.path.exists(dst):
+                    shutil.copyfile(src, dst)
 
         except Exception as e:
             msg = "{0}"
